+++ mxsniff/__init__.py
@@ -17,12 +17,78 @@ from .providers import providers
 
 __all__ = ['MXLookupException', 'get_domain', 'mxsniff', 'mxbulksniff']
 
+_value = ()  # Used in WildcardDomainDict as a placeholder
 
-provider_domains = {}
+
+class WildcardDomainDict(object):
+    """
+    Like a dict, but with custom __getitem__ and __setitem__ to make a nested dictionary
+    with wildcard support for domain name mappings.
+
+    >>> d = WildcardDomainDict()
+    >>> d
+    WildcardDomainDict({})
+    >>> d['*.example.com'] = 'example-wildcard'
+    >>> d['*.wildcard.example.com.'] = 'example-subdotted'
+    >>> d['example.com'] = 'example'
+    >>> d['www.example.com'] = 'example-www'
+    >>> d['example.com']
+    'example'
+    >>> d['www.example.com']
+    'example-www'
+    >>> d['wildcard.example.com']
+    'example-wildcard'
+    >>> d['sub.wildcard.example.com']
+    'example-subdotted'
+    """
+    def __init__(self, *args, **kwargs):
+        self.tree = dict(*args, **kwargs)
+
+    def __repr__(self):
+        return self.__class__.__name__ + '(' + repr(self.tree) + ')'
+
+    def _makeparts(self, key):
+        parts = key.lower().split('.')
+        while '' in parts:
+            parts.remove('')  # Handle trailing dot
+        return parts[::-1]
+
+    def __setitem__(self, key, value):
+        parts = self._makeparts(key)
+        tree = self.tree
+        for item in parts:
+            if item not in tree:
+                tree[item] = {}
+            tree = tree[item]
+        tree[_value] = value
+
+    def __getitem__(self, key):
+        parts = self._makeparts(key)
+        length = len(parts)
+        tree = self.tree
+        for counter, item in enumerate(parts):
+            last = counter == length - 1
+            if item in tree and (not last or last and _value in tree[item]):
+                tree = tree[item]
+            elif '*' in tree:
+                tree = tree['*']
+        if _value in tree:
+            return tree[_value]
+        else:
+            raise KeyError(key)
+
+    def get(self, key, default=None):
+        try:
+            return self[key]
+        except KeyError:
+            return default
+
+
+provider_domains = WildcardDomainDict()
 
 for name, data in providers.items():
     for domain in data['mx']:
-        provider_domains[domain.lower()] = name
+        provider_domains[domain] = name
 
 
 class MXLookupException(Exception):
@@ -60,7 +126,9 @@ def mxsniff(email_or_domain, verbose=False, ignore_errors=False):
     :return: Identified service provider, or a list if there's more than one (in unusual circumstances)
 
     >>> mxsniff('example.com')
+    'nomx'
     >>> mxsniff('__invalid_domain_name__.com')
+    'nomx'
     >>> mxsniff('example@gmail.com')
     'google-gmail'
     >>> mxsniff('https://google.com/')
@@ -77,10 +145,9 @@ def mxsniff(email_or_domain, verbose=False, ignore_errors=False):
         answers = sorted([(rdata.preference, rdata.exchange.to_text(omit_final_dot=True).lower())
             for rdata in dns.resolver.query(domain, 'MX')])
         for preference, exchange in answers:
-            if exchange in provider_domains:
-                provider = provider_domains[exchange]
-                if provider not in result:
-                    result.append(provider)
+            provider = provider_domains.get(exchange)
+            if provider and provider not in result:
+                result.append(provider)
     except (dns.resolver.NoAnswer, dns.resolver.NXDOMAIN, dns.resolver.NoNameservers):
         pass
     except dns.exception.DNSException as e:
@@ -90,12 +157,23 @@ def mxsniff(email_or_domain, verbose=False, ignore_errors=False):
             raise MXLookupException('{exc} {error} ({domain})'.format(
                 exc=e.__class__.__name__, error=text_type(e), domain=domain))
 
+    if not result:
+        # Check for self-hosted email servers, identified with the label 'self'
+        rdomain = tldextract.extract(domain).registered_domain
+        for preference, exchange in answers:
+            if tldextract.extract(exchange).registered_domain == rdomain:
+                result.append('self')
+                break
+        if not result:
+            if answers:
+                result.append('unknown')  # We don't know this one's provider
+            else:
+                result.append('nomx')  # This domain has no mail servers
+
     if verbose:
         return {'name': email_or_domain, 'match': result, 'mx': answers}
     else:
-        if len(result) == 0:
-            return None
-        elif len(result) == 1:
+        if len(result) == 1:
             return result[0]
         else:
             return result
@@ -107,7 +185,7 @@ def mxbulksniff(items, verbose=False, ignore_errors=True):
     repeat queries. Returns a generator that yields one item at a time
 
     >>> list(mxbulksniff(['example.com', 'google.com', 'http://www.google.com']))
-    [('example.com', None), ('google.com', 'google-apps'), ('http://www.google.com', 'google-apps')]
+    [('example.com', 'nomx'), ('google.com', 'google-apps'), ('http://www.google.com', 'google-apps')]
     """
     domain_cache = {}
     for i in items: