+++ CHANGES.rst
@@ -1,6 +1,10 @@
 0.2.2
 =====
 
+* Added support for wildcards in domain names
+* Additional providers
+* Detect self-hosted email servers
+
 0.2.1
 =====